+++ pages/3 Simulate.py
@@ -1,16 +1,45 @@
 import streamlit as st
 import hashlib
 import io
+import ast
 import matplotlib as plt
 import matplotlib
-from plots import get_stock_data
-from simulation import simulate_trading, optimize_choice, simulate_buy_hold
+from functions.plots import get_stock_data
+from functions.simulation import simulate_trading, optimize_choice, simulate_buy_hold
 
 st.set_page_config(layout="wide",
                    page_title='Stock Simulation')
 
 
-def draw_data(column, stats, choice, plot):
+def handle_ARIMA_params(params) -> tuple:
+    components = []
+    temp = ''
+    in_list = False
+    for char in params:
+        if char == '[':
+            in_list = True
+        elif char == ']':
+            in_list = False
+        if char == ',' and not in_list:
+            components.append(temp)
+            temp = ''
+        else:
+            temp += char
+    if temp:
+        components.append(temp)
+    # Process each component
+    tuple_elements = []
+    for component in components:
+        if component.startswith('[') and component.endswith(']'):
+            # If the component is a list representation
+            tuple_elements.append(ast.literal_eval(component))
+        else:
+            # Otherwise, treat it as an integer
+            tuple_elements.append(int(component))
+    return tuple(tuple_elements)
+
+
+def draw_data(column, stats, choice, plot) -> None:
     with column:
         st.write('')
         st.write('')
@@ -31,12 +60,12 @@ def draw_data(column, stats, choice, plot):
             st.write('Money will return after `1` day when selling')
             st.write('Expected market return: `15%`')
             st.write('Risk-free rate: `4.5%`')
-            st.write(f'Position sizing: `{st.session_state["position_sizing"]:.2f}%`')
+            st.write(f'Position sizing: `{st.session_state["position_sizing"]*100:.2f}%`')
 
 
 
 if (('symbol' and 'days_away') not in st.session_state) or not st.session_state['symbol']:
-    st.write('Please head on over to Analysis and pick a symbol (or days away)')
+    st.write('Please head on over to `Analyze` and pick a symbol (or days away)')
 else:
     col1, col2 = st.columns(2)
     with col1:
@@ -47,10 +76,8 @@ else:
             choice = st.selectbox('Pick a Strategy', options=["Random", "Momentum", 'Mean Reversion', 'ARIMA'], label_visibility='collapsed', placeholder='Pick a Strategy', index=None)
             st.session_state['choice'] = choice
             period = st.slider('Pick a lookback period', min_value=1, max_value=30) if (choice == 'Momentum' or choice == 'Mean Reversion') else None
-            params = st.text_input('Pick ARIMA parameters (p,d,q)', placeholder="p,d,q").split(',') if choice == 'ARIMA' else '0,0,0'
-            order = (tuple(map(int, filter(None, params)))
-                     if params != '0,0,0'
-                     else None)
+            params = st.text_input('Pick ARIMA parameters (p,d,q)', placeholder="p,d,q") if choice == 'ARIMA' else '0,0,0'
+            order = handle_ARIMA_params(params)
 
         with col2s:
             Simulate = st.button('Simulate')
