+++ requirements.txt
@@ -52,14 +52,14 @@ referencing==0.33.0
 requests==2.31.0
 rich==13.7.0
 rpds-py==0.18.0
-scipy==1.10.1
+scipy==1.12.0
 six==1.16.0
 smmap==5.0.1
 sniffio==1.3.0
 soupsieve==2.5
 SQLAlchemy==2.0.27
 statsmodels==0.14.1
-streamlit==1.31.1
+streamlit==1.32.0
 tenacity==8.2.3
 toml==0.10.2
 toolz==0.12.1
